Refactor PasswordResets page
