+++ src/pages/PasswordResets.tsx
@@ -1,19 +1,11 @@
-import {
-  Button,
-  FloatingLabel,
-  FormControl,
-  Modal,
-  ModalBody,
-  ModalFooter,
-} from 'react-bootstrap'
 import icon from '../images/icon.svg'
 import BasicPage from 'src/components/BasicPage'
-import { useForm } from 'react-hook-form'
-import { Link, useNavigate, useParams } from 'react-router-dom'
+import { useNavigate, useParams } from 'react-router-dom'
 import { useEffect, useState } from 'react'
 import { ApiPasswordResets } from 'src/api'
 import toast from 'react-hot-toast'
-import { ErrorMessage } from '@hookform/error-message'
+import PasswordResetsForm from 'src/components/PasswordResets/PasswordResetsForm'
+import SuccessModal from 'src/components/PasswordResets/SuccessModal'
 
 interface IPasswordResetsFormState {
   newPassword: string
@@ -24,19 +16,6 @@ export default function PasswordResets() {
   const { token } = useParams()
   const [success, setSuccess] = useState(false)
   const navigate = useNavigate()
-  const {
-    register,
-    watch,
-    getFieldState,
-    handleSubmit,
-    formState: { errors, isValid, isDirty },
-  } = useForm({
-    mode: 'onTouched',
-    defaultValues: {
-      newPassword: '',
-      newPasswordConfirm: '',
-    },
-  })
 
   useEffect(() => {
     async function validateToken() {
@@ -51,19 +30,7 @@ export default function PasswordResets() {
     validateToken()
   }, [token, navigate])
 
-  function validationProps(fieldName: 'newPassword' | 'newPasswordConfirm'): {
-    isValid: boolean
-    isInvalid: boolean
-  } {
-    const { isTouched, invalid, error } = getFieldState(fieldName)
-
-    return {
-      isValid: isTouched && !invalid,
-      isInvalid: isTouched && !!error,
-    }
-  }
-
-  const onSubmit = async ({
+  const updatePassword = async ({
     newPassword,
     newPasswordConfirm,
   }: IPasswordResetsFormState) => {
@@ -84,65 +51,8 @@ export default function PasswordResets() {
 
   return (
     <BasicPage icon={icon} title="New password" fullScreen>
-      <form onSubmit={handleSubmit(onSubmit)} className="vertical-rhythm px-2">
-        <FloatingLabel label="New Password">
-          <FormControl
-            type="password"
-            placeholder="SecurePassword123"
-            {...validationProps('newPassword')}
-            autoComplete="false"
-            {...register('newPassword', {
-              required: 'Password is required',
-              minLength: {
-                value: 8,
-                message: 'Password must be at least 8 characters long',
-              },
-            })}
-          />
-          <FormControl.Feedback type="invalid">
-            <ErrorMessage errors={errors} name="newPassword" />
-          </FormControl.Feedback>
-        </FloatingLabel>
-        <FloatingLabel label="Confirm New Password">
-          <FormControl
-            type="password"
-            placeholder="SecurePassword12334"
-            autoComplete="false"
-            {...validationProps('newPasswordConfirm')}
-            {...register('newPasswordConfirm', {
-              required: 'Password confirmation is required',
-              validate: (value) =>
-                value === watch('newPassword') || 'Passwords do not match',
-            })}
-          />
-          <FormControl.Feedback type="invalid">
-            <ErrorMessage errors={errors} name="newPasswordConfirm" />
-          </FormControl.Feedback>
-        </FloatingLabel>
-        <div className="d-flex flex-row justify-content-between">
-          <Link className="col btn btn-tertiary btn-lg me-3" to="/">
-            Cancel
-          </Link>
-          <Button type="submit" size="lg" disabled={isDirty && !isValid}>
-            Update password
-          </Button>
-        </div>
-      </form>
-      <Modal show={success} centered>
-        <ModalBody className="py-5 text-center">
-          <i className="bi bi-check-circle text-success fs-1" />
-          <h3 className="text-center">Password change successful</h3>
-        </ModalBody>
-        <ModalFooter className="border-0">
-          <Link
-            className="col btn btn-primary btn-lg me-3"
-            to="/"
-            replace={true}
-          >
-            Please sign in
-          </Link>
-        </ModalFooter>
-      </Modal>
+      <PasswordResetsForm onSubmit={updatePassword} />
+      <SuccessModal show={success} />
     </BasicPage>
   )
 }
